requirements.txt
fastapi
uvicorn
pandas
openpyxl
aiosmtplib
python-dotenv 
python-multipart